Refactor script to add full code analysis in a single file
